Align declarative GraphQL scalar API names
