+++ declarative/codegen/src/main/java/io/helidon/declarative/codegen/graphql/server/GraphQlServerCodegenTypes.java
@@ -29,8 +29,8 @@ final class GraphQlServerCodegenTypes {
     static final TypeName GRAPHQL_NON_NULL = TypeName.create("io.helidon.graphql.GraphQl.NonNull");
     static final TypeName GRAPHQL_IGNORE = TypeName.create("io.helidon.graphql.GraphQl.Ignore");
     static final TypeName GRAPHQL_SCALAR = TypeName.create("io.helidon.graphql.GraphQl.Scalar");
-    static final TypeName GRAPHQL_CUSTOM_SCALAR_SPI = TypeName.create("io.helidon.graphql.spi.CustomScalar");
-    static final TypeName GRAPHQL_SCALAR_SPI = TypeName.create("io.helidon.graphql.spi.GraphQlScalar");
+    static final TypeName GRAPHQL_CUSTOM_SCALAR_SPI = TypeName.create("io.helidon.graphql.spi.GraphQlScalar");
+    static final TypeName GRAPHQL_SCALAR_SPI = TypeName.create("io.helidon.graphql.GeneratedGraphQl.CustomScalar");
 
     static final TypeName COMMON_CONTEXT = TypeName.create("io.helidon.common.context.Context");
     static final TypeName COMMON_TYPE_NAME = TypeName.create("io.helidon.common.types.TypeName");
