+++ graphql/graphql/src/main/java/io/helidon/graphql/spi/GraphQlScalar.java
@@ -16,44 +16,21 @@
 
 package io.helidon.graphql.spi;
 
-import java.util.Objects;
-
 import io.helidon.common.Api;
 import io.helidon.service.registry.Service;
 
 /**
- * Service Registry contract for GraphQL custom scalar mappings.
+ * Service Registry contract for declarative GraphQL custom scalar coercion.
  * <p>
- * Implementations can be provided by generated code or by applications. Server and client integrations use this contract
- * to convert between Java values and GraphQL scalar values.
+ * Declarative integrations generate the matching {@link io.helidon.graphql.GeneratedGraphQl.CustomScalar} adapter from the
+ * annotated scalar Java type and delegate conversion to this contract.
+ *
+ * @param <T> Java scalar type
  */
 @Api.Preview
 @Api.Since("27.0.0")
 @Service.Contract
-public interface GraphQlScalar {
-    /**
-     * GraphQL scalar name.
-     *
-     * @return scalar name
-     */
-    String name();
-
-    /**
-     * Java type represented by this scalar.
-     *
-     * @return Java scalar type
-     */
-    Class<?> type();
-
-    /**
-     * Description of this scalar.
-     *
-     * @return scalar description, or an empty string
-     */
-    default String description() {
-        return "";
-    }
-
+public interface GraphQlScalar<T> {
     /**
      * Convert a Java value to a GraphQL scalar result value.
      * <p>
@@ -62,7 +39,7 @@ public interface GraphQlScalar {
      * @param value Java value, never {@code null}
      * @return GraphQL scalar value, never {@code null}
      */
-    Object serialize(Object value);
+    Object serialize(T value);
 
     /**
      * Convert a GraphQL variable value to a Java value.
@@ -72,19 +49,6 @@ public interface GraphQlScalar {
      * @param value GraphQL input value, never {@code null}
      * @return Java value, never {@code null}
      */
-    Object parseValue(Object value);
+    T parseValue(Object value);
 
-    /**
-     * Convert a GraphQL literal value to a Java value.
-     * <p>
-     * GraphQL integrations handle {@code null} values before invoking this method. String, integer, floating point,
-     * boolean, and enum literals are provided as scalar Java values. List literals are provided as {@link java.util.List}
-     * values and input object literals are provided as {@link java.util.Map} values.
-     *
-     * @param value GraphQL literal value converted to a Java value, never {@code null}
-     * @return Java value, never {@code null}
-     */
-    default Object parseLiteral(Object value) {
-        return Objects.requireNonNull(parseValue(Objects.requireNonNull(value)), "parseLiteral result");
-    }
 }
